smart-ingestor/requirements.txt
flask
httpx[http2]
pandas
tqdm
kafka-python
backoff